# [Chore] Webpack optimization [MER-1512]