… assets/yarn.lock