This PR tweaks our webpack config for more efficient builds. Some results:

| Measurement |  Before PR | After PR |
| --------------- | ------------ | --------- |
| Production Build Time | 91 seconds | 88 seconds |
| Development incremental Build Time after modifying JS file | 10s | 2s |
| Development incremental Build Time after modifying scss file | 10s | 10s |
| Total JS File size in build dir | 118M | 31M |
| Total js resources loaded on core-authoring page | 76.9M | 30.5 |
| Total js resources loaded on development version of core-authoring page | 176M | 36M |
| Total js resources loaded on development version of adaptive-authoring page | 57M | 41M |
| Total js resources loaded on adaptive-authoring page | 25.8M | 23M |
| Total js resources loaded on preview of a core-authoring page | 76.9 | 27.9 |
| Total js resources loaded on preview of an adaptive-authoring page | 14.5M | 16.9M |
| Total JS resources on delivery-mode of the of a core authoring page.| 29.6M | 21.2M |
| Total JS resources on delivery-mode of the Blue Planet keeping balance adaptive lesson | 17.4M | 19.8M |


Two lines stand out there as being worse than before, namely the preview & delivery of adaptive lessons.  Those numbers hide the benefit that now, all our vendor code is shared on all pages, which accounts for 11M of that figure. If we assume these learners will also hit another page (sign in, the section list, etc.) than a bulk of that loading will already be done and cached in their browser and they'll have a better overall experience.

I kept a detailed log of everything I tried in the comment section of
https://eliterate.atlassian.net/browse/MER-1512

There are more ideal solutions, but it would require bigger changes and more vigilance in our development going forward because we have so many separate entry-points and code is shared heavily between those entry points.